remove show_versions from top level
