+++ .github/ISSUE_TEMPLATE/bug_report.yml
@@ -80,7 +80,7 @@ body:
       Please run the following code snippet and paste the output here.
 
       ```python
-      from aeon import show_versions; show_versions()
+      from aeon.utils import show_versions; show_versions()
       ```
     placeholder: |
       <details>
