+++ aeon/__init__.py
@@ -1,7 +1,3 @@
 """aeon toolkit."""
 
 __version__ = "0.11.1"
-
-__all__ = ["show_versions"]
-
-from aeon.utils._maint._show_versions import show_versions
