+++ aeon/utils/__init__.py
@@ -9,6 +9,8 @@ __all__ = [
     "COLLECTIONS_DATA_TYPES",
     "SERIES_DATA_TYPES",
     "HIERARCHICAL_DATA_TYPES",
+    # github debug util
+    "show_versions",
 ]
 
 from aeon.utils._data_types import (
@@ -17,5 +19,6 @@ from aeon.utils._data_types import (
     HIERARCHICAL_DATA_TYPES,
     SERIES_DATA_TYPES,
 )
+from aeon.utils._show_versions import show_versions
 from aeon.utils._split import split_series
 from aeon.utils.index_functions import get_cutoff, get_window, update_data
